% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{lower.quartile}
\alias{lower.quartile}
\title{lower.quartile}
\usage{
lower.quartile(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{a vector}

\item{na.rm}{a logical value specifying whether missing values should be removed from the calculations specified by the.functions.}

\item{...}{additional arguments to be passed}
}
\value{
Returns a numeric value representing the lower quartile (25th percentile) of the given vector 'x'.
The lower quartile is the value below which 25% of the data points in the sorted vector fall.
If 'na.rm' is TRUE, any missing values (NA) are removed before the calculation.
The function utilizes the 'quantile' function internally to compute this statistic, and any additional arguments provided (...) are passed along to this underlying function.
}
\description{
value that cuts off the first 25% of the data when it is sorted in ascending order.
}
