% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{max.numerics}
\alias{max.numerics}
\title{max.numerics}
\usage{
\method{max}{numerics}(x, na.rm = TRUE, non.numeric.value = "missing", ...)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{A logical value specifying whether missing values should be removed from the calculations specified by the function. Defaults to TRUE.}

\item{non.numeric.value}{If "missing", returns NA for variables that are not numeric, integer, logical, or complex. Otherwise, returns the first entry of the vector. Defaults to "missing".}

\item{...}{Additional arguments .}
}
\value{
If x is numeric, integer, logical, or complex, the maximal value will be computed. Otherwise, the first value of x will be returned untouched or NA based on non.numeric.value.
}
\description{
This function computes the maximal value of a numeric, integer, logical, or complex vector. If the vector is not one of these types, the function returns either NA or the first entry of the vector.
}
