% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{mean.measured}
\alias{mean.measured}
\title{mean.measured}
\usage{
\method{mean}{measured}(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional arguments.}
}
\value{
A numeric value representing the proportion of non-NA values in the vector.
}
\description{
This function calculates the proportion of non-NA values in a vector.
}
