% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{mean.missing}
\alias{mean.missing}
\title{Calculate the Proportion of NA Values}
\usage{
\method{mean}{missing}(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional arguments.}
}
\value{
A numeric value representing the proportion of NA values in the vector x.
}
\description{
This function calculates the proportion of NA values in a vector.
}
