% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{quantile.numerics}
\alias{quantile.numerics}
\title{This internal function calculates the quantiles for a numeric or complex vector. If the input vector is not numeric or complex, it returns a vector of NA of the same length as the provided probabilities.}
\usage{
\method{quantile}{numerics}(x, probs, na.rm = TRUE)
}
\arguments{
\item{x}{A vector to calculate quantiles for.}

\item{probs}{A numeric vector of probabilities with values in [0,1].}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds. Defaults toTRUE.}
}
\value{
A vector of quantiles corresponding to the specified probabilities if x is numeric or complex. If x is not numeric or complex, returns a vector of NA.
}
\description{
This internal function calculates the quantiles for a numeric or complex vector. If the input vector is not numeric or complex, it returns a vector of NA of the same length as the provided probabilities.
}
\keyword{internal}
