% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{round.exactly.one.value}
\alias{round.exactly.one.value}
\title{round.exactly.one.value}
\usage{
\method{round}{exactly.one.value}(x, digits)
}
\arguments{
\item{x}{A character vector where the first element is the integer part and the second element is the decimal part.}

\item{digits}{The number of digits to round to.}
}
\value{
A character string representing the rounded value.
}
\description{
This internal function rounds exactly one value to the specified number of digits.
}
