% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{round.numerics}
\alias{round.numerics}
\title{round.numerics}
\usage{
\method{round}{numerics}(x, digits = 0, ...)
}
\arguments{
\item{x}{A vector.}

\item{digits}{The number of digits to round to. Defaults to 0.}

\item{...}{Additional arguments.}
}
\value{
If x is numeric or complex, the values will be rounded to the specified number of digits. Otherwise, the values of x will be returned untouched.
}
\description{
This function rounds numeric or complex values to the specified number of digits. If the input is not numeric or complex, the values are returned untouched.
}
