% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{round_exactly}
\alias{round_exactly}
\title{round_exactly}
\usage{
round_exactly(x, digits = 0, decimal = ".", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{The number of digits to round to. This number will be exact, meaning there will be exactly this number of decimal places listed even if this includes trailing zeros. For instance, setting digits = 5 for x = 2.54 would result in 2.54000.}

\item{decimal}{The character specifying the decimal, which splits between whole numbers and the fractional component. Defaults to ".".}

\item{...}{Additional arguments .}
}
\value{
A character vector of rounded numeric values with exactly the specified number of decimal places.
}
\description{
This function rounds numeric values to a specified number of decimal places. The rounding is exact, meaning there will be exactly the specified number of decimal places even if this includes trailing zeros.
}
