% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{round_exactly_one_value}
\alias{round_exactly_one_value}
\title{round_exactly_one_value
This internal function rounds exactly one value to the specified number of digits.}
\usage{
round_exactly_one_value(x, digits)
}
\arguments{
\item{x}{A character vector where the first element is the integer part and the second element is the decimal part.}

\item{digits}{The number of digits to round to.}
}
\value{
A character string representing the rounded value.
}
\description{
round_exactly_one_value
This internal function rounds exactly one value to the specified number of digits.
}
