% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{total.measured}
\alias{total.measured}
\title{total.measured}
\usage{
total.measured(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional arguments.}
}
\value{
An integer representing the total number of non-NA values in the vector x.
}
\description{
This function calculates the total number of non-NA values in a vector.
}
