% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTWrapper2.R
\name{total.missing}
\alias{total.missing}
\title{total.missing}
\usage{
total.missing(x, ...)
}
\arguments{
\item{x}{A vector.}

\item{...}{Additional arguments.}
}
\value{
An integer representing the total number of NA values in the vector x.
}
\description{
This function calculates the total number of NA values in a vector.
}
